requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.3
numpy==1.26.2

# PDF Generation
reportlab==4.0.7
Pillow==10.1.0

# API Requests
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Production
python-dotenv==1.0.0